+++ packages/spotlight/src/spotlight.js
@@ -1276,6 +1276,21 @@ const Spotlight = (function() {
 		},
 
 		/**
+		 * Gets the muted mode value of a spottable element.
+		 *
+		 * @param {Object} [elem] The dom element used to determine the muted status.
+		 * @return {Boolean} `true` if the passed-in control is in muted mode.
+		 * @public
+		 */
+		isMuted: function(elem) {
+			if (!elem) {
+				return false;
+			}
+
+			return matchSelector(elem, '[data-container-muted="true"] .' + spottableClass);
+		},
+
+		/**
 		 * Determines whether Spotlight is currently paused.
 		 *
 		 * @return {Boolean} `true` if Spotlight is currently paused.